requirements.txt
langchain-core==0.3.40
openai==1.74.0
# streamlit==1.41.1
email_validator==2.2.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic_core==2.20.1
pydantic-settings==2.7.0
backoff==2.2.1
fastapi==0.114.1
uvicorn==0.30.6
streamlit-ace==0.1.1
streamlit-extras==0.5.0
boto3==1.37.18
botocore==1.37.18
httpx==0.27.0
st-theme==1.2.3
instructor==1.7.9
imgkit==1.2.3
pydub==0.25.1
pypdf==5.1.0
streamlit-pdf-viewer==0.0.19
pdf2image==1.17.0
unidecode==1.3.8
seaborn==0.13.2
aiosqlite==0.21.0
google-auth==2.38.0
pyasn1-modules==0.4.1
apscheduler==3.11.0
httptools==0.6.4
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
python-multipart==0.0.20
bugsnag==4.7.1
arize-phoenix==10.10.0
arize-phoenix-otel==0.10.3
openinference-instrumentation-openai==0.1.30